+++ integration/integration_test.go
@@ -54,6 +54,7 @@ import (
 	apievents "github.com/gravitational/teleport/api/types/events"
 	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/teleport/api/utils/keypaths"
+	"github.com/gravitational/teleport/integration/helpers"
 	"github.com/gravitational/teleport/lib"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/auth/testauthority"
@@ -75,45 +76,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-const (
-	HostID = "00000000-0000-0000-0000-000000000000"
-	Site   = "local-site"
-)
-
 type integrationTestSuite struct {
-	me *user.User
-	// priv/pub pair to avoid re-generating it
-	priv []byte
-	pub  []byte
-	// log defines the test-specific logger
-	log utils.Logger
+	helpers.Fixture
 }
 
 func newSuite(t *testing.T) *integrationTestSuite {
-	suite := &integrationTestSuite{}
-
-	var err error
-	suite.priv, suite.pub, err = testauthority.New().GenerateKeyPair()
-	require.NoError(t, err)
-
-	// Find AllocatePortsNum free listening ports to use.
-	suite.me, _ = user.Current()
-
-	// close & re-open stdin because 'go test' runs with os.stdin connected to /dev/null
-	stdin, err := os.Open("/dev/tty")
-	if err == nil {
-		os.Stdin.Close()
-		os.Stdin = stdin
-	}
-
-	t.Cleanup(func() {
-		// restore os.Stdin to its original condition: connected to /dev/null
-		os.Stdin.Close()
-		os.Stdin, err = os.Open("/dev/null")
-		require.NoError(t, err)
-	})
-
-	return suite
+	return &integrationTestSuite{*helpers.NewFixture(t)}
 }
 
 type integrationTest func(t *testing.T, suite *integrationTestSuite)
@@ -123,36 +91,11 @@ func (s *integrationTestSuite) bind(test integrationTest) func(t *testing.T) {
 		// Attempt to set a logger for the test. Be warned that parts of the
 		// Teleport codebase do not honour the logger passed in via config and
 		// will create their own. Do not expect to catch _all_ output with this.
-		s.log = utils.NewLoggerForTests()
+		s.Log = utils.NewLoggerForTests()
 		os.RemoveAll(profile.FullProfilePath(""))
-		t.Cleanup(func() { s.log = nil })
+		t.Cleanup(func() { s.Log = nil })
 		test(t, s)
 	}
-}
-
-// newTeleportWithConfig is a helper function that will create a running
-// Teleport instance with the passed in user, instance secrets, and Teleport
-// configuration.
-func (s *integrationTestSuite) newTeleportWithConfig(t *testing.T, logins []string, instanceSecrets []*InstanceSecrets, teleportConfig *service.Config) *TeleInstance {
-	teleport := s.newTeleportInstance()
-
-	// use passed logins, but use suite's default login if nothing was passed
-	if len(logins) == 0 {
-		logins = []string{s.me.Username}
-	}
-	for _, login := range logins {
-		teleport.AddUser(login, []string{login})
-	}
-
-	// create a new teleport instance with passed in configuration
-	if err := teleport.CreateEx(t, instanceSecrets, teleportConfig); err != nil {
-		t.Fatalf("Unexpected response from CreateEx: %v", trace.DebugReport(err))
-	}
-	if err := teleport.Start(); err != nil {
-		t.Fatalf("Unexpected response from Start: %v", trace.DebugReport(err))
-	}
-
-	return teleport
 }
 
 // TestIntegrations acts as the master test suite for all integration tests
@@ -224,26 +167,26 @@ func testDifferentPinnedIP(t *testing.T, suite *integrationTestSuite) {
 	tconf.SSH.Enabled = true
 	tconf.SSH.DisableCreateHostUser = true
 
-	teleport := suite.newTeleportInstance()
+	teleport := suite.NewTeleportInstance()
 
 	role := services.NewImplicitRole()
 	ro := role.GetOptions()
 	ro.PinSourceIP = true
 	role.SetOptions(ro)
 	role.SetName("x")
-	teleport.AddUserWithRole(suite.me.Username, role)
+	teleport.AddUserWithRole(suite.Me.Username, role)
 
 	require.NoError(t, teleport.CreateEx(t, nil, tconf))
 	require.NoError(t, teleport.Start())
 	defer teleport.StopAll()
 
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	for _, ip := range []string{"1.2.3.4/32", "1843:4545::12/128"} {
-		cl, err := teleport.NewClient(ClientConfig{
-			Login:    suite.me.Username,
-			Cluster:  Site,
+		cl, err := teleport.NewClient(helpers.ClientConfig{
+			Login:    suite.Me.Username,
+			Cluster:  helpers.Site,
 			Host:     Host,
 			SourceIP: ip,
 		})
@@ -301,7 +244,7 @@ func testAuditOn(t *testing.T, suite *integrationTestSuite) {
 
 	for _, tt := range tests {
 		t.Run(tt.comment, func(t *testing.T) {
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				auditConfig, err := types.NewClusterAuditConfig(types.ClusterAuditConfigSpecV2{
 					AuditSessionsURI: tt.auditSessionsURI,
 				})
@@ -322,11 +265,11 @@ func testAuditOn(t *testing.T, suite *integrationTestSuite) {
 				tconf.SSH.Enabled = true
 				return t, nil, nil, tconf
 			}
-			teleport := suite.newTeleportWithConfig(makeConfig())
+			teleport := suite.NewTeleportWithConfig(makeConfig())
 			defer teleport.StopAll()
 
 			// Start a node.
-			nodeSSHPort := ports.PopInt()
+			nodeSSHPort := helpers.NewPortValue()
 			nodeConfig := func() *service.Config {
 				tconf := suite.defaultServiceConfig()
 
@@ -342,7 +285,7 @@ func testAuditOn(t *testing.T, suite *integrationTestSuite) {
 			require.NoError(t, err)
 
 			// get access to a authClient for the cluster
-			site := teleport.GetSiteAPI(Site)
+			site := teleport.GetSiteAPI(helpers.Site)
 			require.NotNil(t, site)
 
 			// wait 10 seconds for both nodes to show up, otherwise
@@ -377,9 +320,9 @@ func testAuditOn(t *testing.T, suite *integrationTestSuite) {
 			endC := make(chan error)
 			myTerm := NewTerminal(250)
 			go func() {
-				cl, err := teleport.NewClient(ClientConfig{
-					Login:        suite.me.Username,
-					Cluster:      Site,
+				cl, err := teleport.NewClient(helpers.ClientConfig{
+					Login:        suite.Me.Username,
+					Cluster:      helpers.Site,
 					Host:         Host,
 					Port:         nodeSSHPort,
 					ForwardAgent: tt.inForwardAgent,
@@ -413,7 +356,7 @@ func testAuditOn(t *testing.T, suite *integrationTestSuite) {
 				require.NoError(t, err)
 			}
 			// make sure it's us who joined! :)
-			require.Equal(t, suite.me.Username, session.Parties[0].User)
+			require.Equal(t, suite.Me.Username, session.Parties[0].User)
 
 			// lets type "echo hi" followed by "enter" and then "exit" + "enter":
 
@@ -628,9 +571,9 @@ func testInteroperability(t *testing.T, suite *integrationTestSuite) {
 	for i, tt := range tests {
 		t.Run(fmt.Sprintf("Test %d: %s", i, strings.Fields(tt.inCommand)[0]), func(t *testing.T) {
 			// create new teleport client
-			cl, err := teleport.NewClient(ClientConfig{
-				Login:   suite.me.Username,
-				Cluster: Site,
+			cl, err := teleport.NewClient(helpers.ClientConfig{
+				Login:   suite.Me.Username,
+				Cluster: helpers.Site,
 				Host:    Host,
 				Port:    teleport.GetPortSSHInt(),
 			})
@@ -671,7 +614,7 @@ func testInteroperability(t *testing.T, suite *integrationTestSuite) {
 // it as an argument. Otherwise, it will run tests as normal.
 func TestMain(m *testing.M) {
 	utils.InitLoggerForTests()
-	SetTestTimeouts(100 * time.Millisecond)
+	helpers.SetTestTimeouts(100 * time.Millisecond)
 	// If the test is re-executing itself, execute the command that comes over
 	// the pipe.
 	if srv.IsReexec() {
@@ -686,11 +629,11 @@ func TestMain(m *testing.M) {
 
 // newUnstartedTeleport helper returns a created but not started Teleport instance pre-configured
 // with the current user os.user.Current().
-func (s *integrationTestSuite) newUnstartedTeleport(t *testing.T, logins []string, enableSSH bool) *TeleInstance {
-	teleport := s.newTeleportInstance()
+func (s *integrationTestSuite) newUnstartedTeleport(t *testing.T, logins []string, enableSSH bool) *helpers.TeleInstance {
+	teleport := s.NewTeleportInstance()
 	// use passed logins, but use suite's default login if nothing was passed
 	if len(logins) == 0 {
-		logins = []string{s.me.Username}
+		logins = []string{s.Me.Username}
 	}
 	for _, login := range logins {
 		teleport.AddUser(login, []string{login})
@@ -701,7 +644,7 @@ func (s *integrationTestSuite) newUnstartedTeleport(t *testing.T, logins []strin
 
 // newTeleport helper returns a running Teleport instance pre-configured
 // with the current user os.user.Current().
-func (s *integrationTestSuite) newTeleport(t *testing.T, logins []string, enableSSH bool) *TeleInstance {
+func (s *integrationTestSuite) newTeleport(t *testing.T, logins []string, enableSSH bool) *helpers.TeleInstance {
 	teleport := s.newUnstartedTeleport(t, logins, enableSSH)
 	require.NoError(t, teleport.Start())
 	return teleport
@@ -709,7 +652,7 @@ func (s *integrationTestSuite) newTeleport(t *testing.T, logins []string, enable
 
 // newTeleportIoT helper returns a running Teleport instance with Host as a
 // reversetunnel node.
-func (s *integrationTestSuite) newTeleportIoT(t *testing.T, logins []string) *TeleInstance {
+func (s *integrationTestSuite) newTeleportIoT(t *testing.T, logins []string) *helpers.TeleInstance {
 	// Create a Teleport instance with Auth/Proxy.
 	mainConfig := func() *service.Config {
 		tconf := s.defaultServiceConfig()
@@ -723,7 +666,7 @@ func (s *integrationTestSuite) newTeleportIoT(t *testing.T, logins []string) *Te
 
 		return tconf
 	}
-	main := s.newTeleportWithConfig(t, logins, nil, mainConfig())
+	main := s.NewTeleportWithConfig(t, logins, nil, mainConfig())
 
 	// Create a Teleport instance with a Node.
 	nodeConfig := func() *service.Config {
@@ -766,12 +709,12 @@ func testUUIDBasedProxy(t *testing.T, suite *integrationTestSuite) {
 	teleportSvr := suite.newTeleport(t, nil, true)
 	defer teleportSvr.StopAll()
 
-	site := teleportSvr.GetSiteAPI(Site)
+	site := teleportSvr.GetSiteAPI(helpers.Site)
 
 	// addNode adds a node to the teleport instance, returning its uuid.
 	// All nodes added this way have the same hostname.
 	addNode := func() (string, error) {
-		nodeSSHPort := ports.PopInt()
+		nodeSSHPort := helpers.NewPortValue()
 		tconf := suite.defaultServiceConfig()
 		tconf.Hostname = Host
 
@@ -828,14 +771,14 @@ func testUUIDBasedProxy(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 
 	// attempting to run a command by hostname should generate NodeIsAmbiguous error.
-	_, err = runCommand(t, teleportSvr, []string{"echo", "Hello there!"}, ClientConfig{Login: suite.me.Username, Cluster: Site, Host: Host}, 1)
+	_, err = runCommand(t, teleportSvr, []string{"echo", "Hello there!"}, helpers.ClientConfig{Login: suite.Me.Username, Cluster: helpers.Site, Host: Host}, 1)
 	require.Error(t, err)
 	if !strings.Contains(err.Error(), teleport.NodeIsAmbiguous) {
 		require.FailNowf(t, "Expected %s, got %s", teleport.NodeIsAmbiguous, err.Error())
 	}
 
 	// attempting to run a command by uuid should succeed.
-	_, err = runCommand(t, teleportSvr, []string{"echo", "Hello there!"}, ClientConfig{Login: suite.me.Username, Cluster: Site, Host: uuid1}, 1)
+	_, err = runCommand(t, teleportSvr, []string{"echo", "Hello there!"}, helpers.ClientConfig{Login: suite.Me.Username, Cluster: helpers.Site, Host: uuid1}, 1)
 	require.NoError(t, err)
 }
 
@@ -845,13 +788,13 @@ func testSSHTracker(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleport(t, nil, true)
 	defer teleport.StopAll()
 
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	personA := NewTerminal(250)
-	cl, err := teleport.NewClient(ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cl, err := teleport.NewClient(helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Host,
 	})
 	require.NoError(t, err)
@@ -880,7 +823,7 @@ func testInteractiveRegular(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleport(t, nil, true)
 	defer teleport.StopAll()
 
-	verifySessionJoin(t, suite.me.Username, teleport)
+	verifySessionJoin(t, suite.Me.Username, teleport)
 }
 
 // TestInteractiveReverseTunnel covers SSH into shell and joining the same session from another client
@@ -896,7 +839,7 @@ func testInteractiveReverseTunnel(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleportIoT(t, nil)
 	defer teleport.StopAll()
 
-	verifySessionJoin(t, suite.me.Username, teleport)
+	verifySessionJoin(t, suite.Me.Username, teleport)
 }
 
 func testSessionRecordingModes(t *testing.T, suite *integrationTestSuite) {
@@ -917,7 +860,7 @@ func testSessionRecordingModes(t *testing.T, suite *integrationTestSuite) {
 	cfg.Proxy.Enabled = true
 	cfg.SSH.Enabled = true
 
-	teleport := suite.newTeleportWithConfig(t, nil, nil, cfg)
+	teleport := suite.NewTeleportWithConfig(t, nil, nil, cfg)
 	defer teleport.StopAll()
 
 	// startSession starts an interactive session, users must terminate the
@@ -927,9 +870,9 @@ func testSessionRecordingModes(t *testing.T, suite *integrationTestSuite) {
 		errCh := make(chan error)
 
 		go func() {
-			cl, err := teleport.NewClient(ClientConfig{
+			cl, err := teleport.NewClient(helpers.ClientConfig{
 				Login:   username,
-				Cluster: Site,
+				Cluster: helpers.Site,
 				Host:    Host,
 			})
 			if err != nil {
@@ -985,7 +928,7 @@ func testSessionRecordingModes(t *testing.T, suite *integrationTestSuite) {
 	} {
 		t.Run(name, func(t *testing.T) {
 			// Setup user and session recording mode.
-			username := suite.me.Username
+			username := suite.Me.Username
 			role, err := types.NewRoleV3("devs", types.RoleSpecV5{
 				Allow: types.RoleConditions{
 					Logins: []string{username},
@@ -997,7 +940,7 @@ func testSessionRecordingModes(t *testing.T, suite *integrationTestSuite) {
 				},
 			})
 			require.NoError(t, err)
-			require.NoError(t, SetupUser(teleport.Process, username, []types.Role{role}))
+			require.NoError(t, helpers.SetupUser(teleport.Process, username, []types.Role{role}))
 
 			t.Run("BeforeStartFailure", func(t *testing.T) {
 				// Enable disk failure.
@@ -1093,9 +1036,9 @@ func testCustomReverseTunnel(t *testing.T, suite *integrationTestSuite) {
 	}
 	conf.SSH.Enabled = false
 
-	instanceConfig := suite.defaultInstanceConfig()
-	instanceConfig.Ports = webReverseTunnelMuxPortSetup()
-	main := NewInstance(instanceConfig)
+	instanceConfig := suite.DefaultInstanceConfig()
+	instanceConfig.Ports = helpers.WebReverseTunnelMuxPortSetup()
+	main := helpers.NewInstance(instanceConfig)
 
 	require.NoError(t, main.CreateEx(t, nil, conf))
 	require.NoError(t, main.Start())
@@ -1116,9 +1059,9 @@ func testCustomReverseTunnel(t *testing.T, suite *integrationTestSuite) {
 }
 
 // verifySessionJoin covers SSH into shell and joining the same session from another client
-func verifySessionJoin(t *testing.T, username string, teleport *TeleInstance) {
+func verifySessionJoin(t *testing.T, username string, teleport *helpers.TeleInstance) {
 	// get a reference to site obj:
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	personA := NewTerminal(250)
@@ -1127,9 +1070,9 @@ func verifySessionJoin(t *testing.T, username string, teleport *TeleInstance) {
 	// PersonA: SSH into the server, wait one second, then type some commands on stdin:
 	sessionA := make(chan error)
 	openSession := func() {
-		cl, err := teleport.NewClient(ClientConfig{
+		cl, err := teleport.NewClient(helpers.ClientConfig{
 			Login:   username,
-			Cluster: Site,
+			Cluster: helpers.Site,
 			Host:    Host,
 		})
 		if err != nil {
@@ -1155,9 +1098,9 @@ func verifySessionJoin(t *testing.T, username string, teleport *TeleInstance) {
 		}
 
 		sessionID := string(sessions[0].ID)
-		cl, err := teleport.NewClient(ClientConfig{
+		cl, err := teleport.NewClient(helpers.ClientConfig{
 			Login:   username,
-			Cluster: Site,
+			Cluster: helpers.Site,
 			Host:    Host,
 		})
 		if err != nil {
@@ -1211,14 +1154,14 @@ func testShutdown(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleport(t, nil, true)
 
 	// get a reference to site obj:
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	person := NewTerminal(250)
 
-	cl, err := teleport.NewClient(ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cl, err := teleport.NewClient(helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Host,
 		Port:    teleport.GetPortSSHInt(),
 	})
@@ -1314,7 +1257,7 @@ type disconnectTestCase struct {
 	disconnectTimeout time.Duration
 	concurrentConns   int
 	sessCtlTimeout    time.Duration
-	postFunc          func(context.Context, *testing.T, *TeleInstance)
+	postFunc          func(context.Context, *testing.T, *helpers.TeleInstance)
 
 	// verifyError checks if `err` reflects the error expected by the test scenario.
 	// It returns nil if yes, non-nil otherwise.
@@ -1387,8 +1330,8 @@ func testDisconnectScenarios(t *testing.T, suite *integrationTestSuite) {
 			sessCtlTimeout:    500 * time.Millisecond,
 			// use postFunc to wait for the semaphore to be acquired and a session
 			// to be started, then shut down the auth server.
-			postFunc: func(ctx context.Context, t *testing.T, teleport *TeleInstance) {
-				site := teleport.GetSiteAPI(Site)
+			postFunc: func(ctx context.Context, t *testing.T, teleport *helpers.TeleInstance) {
+				site := teleport.GetSiteAPI(helpers.Site)
 				var sems []types.Semaphore
 				var err error
 				for i := 0; i < 6; i++ {
@@ -1426,9 +1369,9 @@ func testDisconnectScenarios(t *testing.T, suite *integrationTestSuite) {
 }
 
 func runDisconnectTest(t *testing.T, suite *integrationTestSuite, tc disconnectTestCase) {
-	teleport := suite.newTeleportInstance()
+	teleport := suite.NewTeleportInstance()
 
-	username := suite.me.Username
+	username := suite.Me.Username
 	role, err := types.NewRoleV3("devs", types.RoleSpecV5{
 		Options: tc.options,
 		Allow: types.RoleConditions{
@@ -1476,9 +1419,9 @@ func runDisconnectTest(t *testing.T, suite *integrationTestSuite, tc disconnectT
 
 		openSession := func() {
 			defer cancel()
-			cl, err := teleport.NewClient(ClientConfig{
+			cl, err := teleport.NewClient(helpers.ClientConfig{
 				Login:   username,
-				Cluster: Site,
+				Cluster: helpers.Site,
 				Host:    Host,
 				Port:    teleport.GetPortSSHInt(),
 			})
@@ -1587,9 +1530,9 @@ func testEnvironmentVariables(t *testing.T, suite *integrationTestSuite) {
 	cmd := []string{"printenv", testKey}
 
 	// make sure sessions set run command
-	tc, err := teleport.NewClient(ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	tc, err := teleport.NewClient(helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Host,
 		Port:    teleport.GetPortSSHInt(),
 	})
@@ -1617,8 +1560,8 @@ func testInvalidLogins(t *testing.T, suite *integrationTestSuite) {
 	cmd := []string{"echo", "success"}
 
 	// try the wrong site:
-	tc, err := teleport.NewClient(ClientConfig{
-		Login:   suite.me.Username,
+	tc, err := teleport.NewClient(helpers.ClientConfig{
+		Login:   suite.Me.Username,
 		Cluster: "wrong-site",
 		Host:    Host,
 		Port:    teleport.GetPortSSHInt(),
@@ -1677,7 +1620,7 @@ func testTwoClustersTunnel(t *testing.T, suite *integrationTestSuite) {
 
 func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time, proxyRecordMode string, execCountSiteA, execCountSiteB int) {
 	// start the http proxy, we need to make sure this was not used
-	ps := &proxyServer{}
+	ps := &helpers.ProxyServer{}
 	ts := httptest.NewServer(ps)
 	defer ts.Close()
 
@@ -1686,7 +1629,7 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 		t.Setenv(v, "")
 	}
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	a := suite.newNamedTeleportInstance(t, "site-A")
 	b := suite.newNamedTeleportInstance(t, "site-B")
@@ -1741,7 +1684,7 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 	cmd := []string{"echo", "hello world"}
 
 	// directly:
-	tc, err := a.NewClient(ClientConfig{
+	tc, err := a.NewClient(helpers.ClientConfig{
 		Login:        username,
 		Cluster:      a.Secrets.SiteName,
 		Host:         Host,
@@ -1785,7 +1728,7 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 	waitForActiveTunnelConnections(t, b.Tunnel, a.Secrets.SiteName, 1)
 
 	// via tunnel b->a:
-	tc, err = b.NewClient(ClientConfig{
+	tc, err = b.NewClient(helpers.ClientConfig{
 		Login:        username,
 		Cluster:      a.Secrets.SiteName,
 		Host:         Host,
@@ -1807,7 +1750,7 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 	require.NoError(t, a.Reset())
 	require.NoError(t, a.Start())
 
-	// try to execute an SSH command using the same old client to Site-B
+	// try to execute an SSH command using the same old client to helpers.Site-B
 	// "site-A" and "site-B" reverse tunnels are supposed to reconnect,
 	// and 'tc' (client) is also supposed to reconnect
 	var sshErr error
@@ -1816,7 +1759,7 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 		return sshErr == nil
 	}
 	require.Eventually(t, tcHasReconnected, 10*time.Second, 250*time.Millisecond,
-		"Timed out waiting for Site A to restart: %v", sshErr)
+		"Timed out waiting for helpers.Site A to restart: %v", sshErr)
 
 	clientHasEvents := func(site auth.ClientI, count int) func() bool {
 		// only look for exec events
@@ -1831,11 +1774,11 @@ func twoClustersTunnel(t *testing.T, suite *integrationTestSuite, now time.Time,
 
 	siteA := a.GetSiteAPI(a.Secrets.SiteName)
 	require.Eventually(t, clientHasEvents(siteA, execCountSiteA), 5*time.Second, 500*time.Millisecond,
-		"Failed to find %d events on Site A after 5s", execCountSiteA)
+		"Failed to find %d events on helpers.Site A after 5s", execCountSiteA)
 
 	siteB := b.GetSiteAPI(b.Secrets.SiteName)
 	require.Eventually(t, clientHasEvents(siteB, execCountSiteB), 5*time.Second, 500*time.Millisecond,
-		"Failed to find %d events on Site B after 5s", execCountSiteB)
+		"Failed to find %d events on helpers.Site B after 5s", execCountSiteB)
 }
 
 // TestTwoClustersProxy checks if the reverse tunnel uses a HTTP PROXY to
@@ -1845,7 +1788,7 @@ func testTwoClustersProxy(t *testing.T, suite *integrationTestSuite) {
 	defer tr.Stop()
 
 	// start the http proxy
-	ps := &proxyServer{}
+	ps := &helpers.ProxyServer{}
 	ts := httptest.NewServer(ps)
 	defer ts.Close()
 
@@ -1854,7 +1797,7 @@ func testTwoClustersProxy(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 	t.Setenv("http_proxy", u.Host)
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	// httpproxy doesn't allow proxying when the target is localhost, so use
 	// this address instead.
@@ -1894,7 +1837,7 @@ func testHA(t *testing.T, suite *integrationTestSuite) {
 	tr := utils.NewTracer(utils.ThisFunction()).Start()
 	defer tr.Stop()
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	a := suite.newNamedTeleportInstance(t, "cluster-a")
 	b := suite.newNamedTeleportInstance(t, "cluster-b")
@@ -1908,7 +1851,7 @@ func testHA(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, b.Start())
 	require.NoError(t, a.Start())
 
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	sshPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
 	require.NoError(t, a.StartNodeAndProxy("cluster-a-node", sshPort, proxyWebPort, proxySSHPort))
 
@@ -1919,7 +1862,7 @@ func testHA(t *testing.T, suite *integrationTestSuite) {
 		"Two clusters do not see each other: tunnels are not working.")
 
 	cmd := []string{"echo", "hello world"}
-	tc, err := b.NewClient(ClientConfig{
+	tc, err := b.NewClient(helpers.ClientConfig{
 		Login:   username,
 		Cluster: "cluster-a",
 		Host:    Loopback,
@@ -1929,7 +1872,7 @@ func testHA(t *testing.T, suite *integrationTestSuite) {
 
 	output := &bytes.Buffer{}
 	tc.Stdout = output
-	// try to execute an SSH command using the same old client  to Site-B
+	// try to execute an SSH command using the same old client  to helpers.Site-B
 	// "site-A" and "site-B" reverse tunnels are supposed to reconnect,
 	// and 'tc' (client) is also supposed to reconnect
 	for i := 0; i < 10; i++ {
@@ -1984,7 +1927,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 	tr := utils.NewTracer(utils.ThisFunction()).Start()
 	defer tr.Stop()
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	clusterMain := "cluster-main"
 	clusterAux := "cluster-aux"
@@ -2004,7 +1947,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 
 	// for role mapping test we turn on Web API on the main cluster
 	// as it's used
-	makeConfig := func(enableSSH bool) (*testing.T, []*InstanceSecrets, *service.Config) {
+	makeConfig := func(enableSSH bool) (*testing.T, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 		tconf.SSH.Enabled = enableSSH
 		tconf.Proxy.DisableWebService = false
@@ -2052,7 +1995,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 	tryCreateTrustedCluster(t, aux.Process.GetAuthServer(), trustedCluster)
 	waitForTunnelConnections(t, main.Process.GetAuthServer(), clusterAux, 1)
 
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	sshPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
 	require.NoError(t, aux.StartNodeAndProxy("aux-node", sshPort, proxyWebPort, proxySSHPort))
 
@@ -2077,7 +2020,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 	require.Len(t, nodes, 2)
 
 	cmd := []string{"echo", "hello world"}
-	tc, err := main.NewClient(ClientConfig{
+	tc, err := main.NewClient(helpers.ClientConfig{
 		Login:   username,
 		Cluster: clusterAux,
 		Host:    Loopback,
@@ -2087,7 +2030,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 	output := &bytes.Buffer{}
 	tc.Stdout = output
 	require.NoError(t, err)
-	// try to execute an SSH command using the same old client  to Site-B
+	// try to execute an SSH command using the same old client  to helpers.Site-B
 	// "site-A" and "site-B" reverse tunnels are supposed to reconnect,
 	// and 'tc' (client) is also supposed to reconnect
 	for i := 0; i < 10; i++ {
@@ -2105,7 +2048,7 @@ func testMapRoles(t *testing.T, suite *integrationTestSuite) {
 		name                       string
 		mainClusterName            string
 		auxClusterName             string
-		inCluster                  *TeleInstance
+		inCluster                  *helpers.TeleInstance
 		outChkMainUserCA           require.ErrorAssertionFunc
 		outChkMainUserCAPrivateKey require.ValueAssertionFunc
 		outChkMainHostCA           require.ErrorAssertionFunc
@@ -2278,26 +2221,26 @@ func testMultiplexingTrustedClusters(t *testing.T, suite *integrationTestSuite) 
 	trustedClusters(t, suite, trustedClusterTest{multiplex: true})
 }
 
-func standardPortsOrMuxSetup(mux bool) *InstancePorts {
+func standardPortsOrMuxSetup(mux bool) *helpers.InstancePorts {
 	if mux {
-		return webReverseTunnelMuxPortSetup()
+		return helpers.WebReverseTunnelMuxPortSetup()
 	}
-	return standardPortSetup()
+	return helpers.StandardPortSetup()
 }
 
 func trustedClusters(t *testing.T, suite *integrationTestSuite, test trustedClusterTest) {
 	ctx := context.Background()
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	clusterMain := "cluster-main"
 	clusterAux := "cluster-aux"
-	main := NewInstance(InstanceConfig{
+	main := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: clusterMain,
-		HostID:      HostID,
+		HostID:      helpers.HostID,
 		NodeName:    Host,
-		Priv:        suite.priv,
-		Pub:         suite.pub,
-		log:         suite.log,
+		Priv:        suite.Priv,
+		Pub:         suite.Pub,
+		Log:         suite.Log,
 		Ports:       standardPortsOrMuxSetup(test.multiplex),
 	})
 	aux := suite.newNamedTeleportInstance(t, clusterAux)
@@ -2341,7 +2284,7 @@ func trustedClusters(t *testing.T, suite *integrationTestSuite, test trustedClus
 
 	// for role mapping test we turn on Web API on the main cluster
 	// as it's used
-	makeConfig := func(enableSSH bool) (*testing.T, []*InstanceSecrets, *service.Config) {
+	makeConfig := func(enableSSH bool) (*testing.T, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 		tconf.Proxy.DisableWebService = false
 		tconf.Proxy.DisableWebInterface = true
@@ -2399,7 +2342,7 @@ func trustedClusters(t *testing.T, suite *integrationTestSuite, test trustedClus
 	tryCreateTrustedCluster(t, aux.Process.GetAuthServer(), trustedCluster)
 	waitForTunnelConnections(t, main.Process.GetAuthServer(), clusterAux, 1)
 
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	sshPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
 	require.NoError(t, aux.StartNodeAndProxy("aux-node", sshPort, proxyWebPort, proxySSHPort))
 
@@ -2413,14 +2356,14 @@ func trustedClusters(t *testing.T, suite *integrationTestSuite, test trustedClus
 
 	// Try and connect to a node in the Aux cluster from the Main cluster using
 	// direct dialing.
-	creds, err := GenerateUserCreds(UserCredsRequest{
+	creds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{
 		Process:        main.Process,
 		Username:       username,
 		RouteToCluster: clusterAux,
 	})
 	require.NoError(t, err)
 
-	tc, err := main.NewClientWithCreds(ClientConfig{
+	tc, err := main.NewClientWithCreds(helpers.ClientConfig{
 		Login:    username,
 		Cluster:  clusterAux,
 		Host:     Loopback,
@@ -2452,7 +2395,7 @@ func trustedClusters(t *testing.T, suite *integrationTestSuite, test trustedClus
 	require.Equal(t, "hello world\n", output.String())
 
 	// Try and generate user creds for Aux cluster as ops user.
-	_, err = GenerateUserCreds(UserCredsRequest{
+	_, err = helpers.GenerateUserCreds(helpers.UserCredsRequest{
 		Process:        main.Process,
 		Username:       mainOps,
 		RouteToCluster: clusterAux,
@@ -2542,7 +2485,7 @@ func waitForClusters(tun reversetunnel.Server, expected int) func() bool {
 
 func testTrustedTunnelNode(t *testing.T, suite *integrationTestSuite) {
 	ctx := context.Background()
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	clusterMain := "cluster-main"
 	clusterAux := "cluster-aux"
@@ -2561,7 +2504,7 @@ func testTrustedTunnelNode(t *testing.T, suite *integrationTestSuite) {
 
 	// for role mapping test we turn on Web API on the main cluster
 	// as it's used
-	makeConfig := func(enableSSH bool) (*testing.T, []*InstanceSecrets, *service.Config) {
+	makeConfig := func(enableSSH bool) (*testing.T, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 		tconf.Proxy.DisableWebService = false
 		tconf.Proxy.DisableWebInterface = true
@@ -2643,7 +2586,7 @@ func testTrustedTunnelNode(t *testing.T, suite *integrationTestSuite) {
 
 	// Try and connect to a node in the Aux cluster from the Main cluster using
 	// direct dialing.
-	tc, err := main.NewClient(ClientConfig{
+	tc, err := main.NewClient(helpers.ClientConfig{
 		Login:   username,
 		Cluster: clusterAux,
 		Host:    Loopback,
@@ -2665,7 +2608,7 @@ func testTrustedTunnelNode(t *testing.T, suite *integrationTestSuite) {
 
 	// Try and connect to a node in the Aux cluster from the Main cluster using
 	// tunnel dialing.
-	tunnelClient, err := main.NewClient(ClientConfig{
+	tunnelClient, err := main.NewClient(helpers.ClientConfig{
 		Login:   username,
 		Cluster: clusterAux,
 		Host:    tunnelNodeHostname,
@@ -2695,10 +2638,10 @@ func testDiscoveryRecovers(t *testing.T, suite *integrationTestSuite) {
 	tr := utils.NewTracer(utils.ThisFunction()).Start()
 	defer tr.Stop()
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	// create load balancer for main cluster proxies
-	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(ports.PopInt())))
+	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, helpers.NewPortStr()))
 	lb, err := utils.NewLoadBalancer(context.TODO(), frontend)
 	require.NoError(t, err)
 	require.NoError(t, lb.Listen())
@@ -2729,11 +2672,11 @@ func testDiscoveryRecovers(t *testing.T, suite *integrationTestSuite) {
 		"Two clusters do not see each other: tunnels are not working.")
 
 	// Helper function for adding a new proxy to "main".
-	addNewMainProxy := func(name string) (reversetunnel.Server, ProxyConfig) {
+	addNewMainProxy := func(name string) (reversetunnel.Server, helpers.ProxyConfig) {
 		t.Logf("adding main proxy %q...", name)
-		nodePorts := ports.PopIntSlice(3)
+		nodePorts := helpers.NewPortSlice(3)
 		proxyReverseTunnelPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
-		newConfig := ProxyConfig{
+		newConfig := helpers.ProxyConfig{
 			Name:              name,
 			SSHPort:           proxySSHPort,
 			WebPort:           proxyWebPort,
@@ -2768,8 +2711,8 @@ func testDiscoveryRecovers(t *testing.T, suite *integrationTestSuite) {
 	// Helper function for testing that a proxy in main has been discovered by
 	// (and is able to use reverse tunnel into) remote.  If conf is nil, main's
 	// first/default proxy will be called.
-	testProxyConn := func(conf *ProxyConfig, shouldFail bool) {
-		clientConf := ClientConfig{
+	testProxyConn := func(conf *helpers.ProxyConfig, shouldFail bool) {
+		clientConf := helpers.ClientConfig{
 			Login:   username,
 			Cluster: "cluster-remote",
 			Host:    Loopback,
@@ -2828,10 +2771,10 @@ func testDiscovery(t *testing.T, suite *integrationTestSuite) {
 	tr := utils.NewTracer(utils.ThisFunction()).Start()
 	defer tr.Stop()
 
-	username := suite.me.Username
+	username := suite.Me.Username
 
 	// create load balancer for main cluster proxies
-	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(ports.PopInt())))
+	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(helpers.NewPortValue())))
 	lb, err := utils.NewLoadBalancer(context.TODO(), frontend)
 	require.NoError(t, err)
 	require.NoError(t, lb.Listen())
@@ -2862,9 +2805,9 @@ func testDiscovery(t *testing.T, suite *integrationTestSuite) {
 		"Two clusters do not see each other: tunnels are not working.")
 
 	// start second proxy
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	proxyReverseTunnelPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
-	proxyConfig := ProxyConfig{
+	proxyConfig := helpers.ProxyConfig{
 		Name:              "cluster-main-proxy",
 		SSHPort:           proxySSHPort,
 		WebPort:           proxyWebPort,
@@ -2883,7 +2826,7 @@ func testDiscovery(t *testing.T, suite *integrationTestSuite) {
 	waitForActiveTunnelConnections(t, secondProxy, "cluster-remote", 1)
 
 	// execute the connection via first proxy
-	cfg := ClientConfig{
+	cfg := helpers.ClientConfig{
 		Login:   username,
 		Cluster: "cluster-remote",
 		Host:    Loopback,
@@ -2897,7 +2840,7 @@ func testDiscovery(t *testing.T, suite *integrationTestSuite) {
 	// tried 10 times with 250 millisecond delay between each attempt to allow
 	// the discovery request to be received and the connection added to the agent
 	// pool.
-	cfgProxy := ClientConfig{
+	cfgProxy := helpers.ClientConfig{
 		Login:   username,
 		Cluster: "cluster-remote",
 		Host:    Loopback,
@@ -2959,7 +2902,7 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 	t.Cleanup(func() { lib.SetInsecureDevMode(false) })
 
 	// Create and start load balancer for proxies.
-	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(ports.PopInt())))
+	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(helpers.NewPortValue())))
 	lb, err := utils.NewLoadBalancer(context.TODO(), frontend)
 	require.NoError(t, err)
 	require.NoError(t, lb.Listen())
@@ -2967,7 +2910,7 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 	t.Cleanup(func() { require.NoError(t, lb.Close()) })
 
 	// Create a Teleport instance with Auth/Proxy.
-	mainConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+	mainConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 
 		tconf.Auth.Enabled = true
@@ -2987,13 +2930,13 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 
 		return t, nil, nil, tconf
 	}
-	main := suite.newTeleportWithConfig(mainConfig())
+	main := suite.NewTeleportWithConfig(mainConfig())
 	t.Cleanup(func() { require.NoError(t, main.StopAll()) })
 
 	// Create a Teleport instance with a Proxy.
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	proxyReverseTunnelPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
-	proxyConfig := ProxyConfig{
+	proxyConfig := helpers.ProxyConfig{
 		Name:                   "cluster-main-proxy",
 		SSHPort:                proxySSHPort,
 		WebPort:                proxyWebPort,
@@ -3031,21 +2974,21 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 
 	// Wait for active tunnel connections to be established.
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 0)
-	waitForActiveTunnelConnections(t, proxyTunnel, Site, 1)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 0)
+	waitForActiveTunnelConnections(t, proxyTunnel, helpers.Site, 1)
 
 	// Execute the connection via first proxy.
-	cfg := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cfg := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    "cluster-main-node",
 	}
 	_, err = runCommand(t, main, []string{"echo", "hello world"}, cfg, 1)
 	require.Error(t, err)
 
-	cfgProxy := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cfgProxy := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    "cluster-main-node",
 		Proxy:   &proxyConfig,
 	}
@@ -3056,8 +2999,8 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 
 	// stop the proxy to collapse the tunnel
 	require.NoError(t, main.StopProxy())
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 0)
-	waitForActiveTunnelConnections(t, proxyTunnel, Site, 0)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 0)
+	waitForActiveTunnelConnections(t, proxyTunnel, helpers.Site, 0)
 
 	// Requests going via both proxy will fail.
 	_, err = runCommand(t, main, []string{"echo", "hello world"}, cfg, 1)
@@ -3079,8 +3022,8 @@ func testReverseTunnelCollapse(t *testing.T, suite *integrationTestSuite) {
 	// start the proxy again and ensure the tunnel is re-established
 	proxyTunnel, err = main.StartProxy(proxyConfig)
 	require.NoError(t, err)
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 0)
-	waitForActiveTunnelConnections(t, proxyTunnel, Site, 1)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 0)
+	waitForActiveTunnelConnections(t, proxyTunnel, helpers.Site, 1)
 
 	// Requests going to the connected proxy should succeed.
 	_, err = runCommand(t, main, []string{"echo", "hello world"}, cfg, 1)
@@ -3103,7 +3046,7 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 	defer lib.SetInsecureDevMode(false)
 
 	// Create and start load balancer for proxies.
-	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(ports.PopInt())))
+	frontend := *utils.MustParseAddr(net.JoinHostPort(Loopback, strconv.Itoa(helpers.NewPortValue())))
 	lb, err := utils.NewLoadBalancer(context.TODO(), frontend)
 	require.NoError(t, err)
 	err = lb.Listen()
@@ -3112,7 +3055,7 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 	defer lb.Close()
 
 	// Create a Teleport instance with Auth/Proxy.
-	mainConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+	mainConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 
 		tconf.Auth.Enabled = true
@@ -3132,13 +3075,13 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 
 		return t, nil, nil, tconf
 	}
-	main := suite.newTeleportWithConfig(mainConfig())
+	main := suite.NewTeleportWithConfig(mainConfig())
 	defer main.StopAll()
 
 	// Create a Teleport instance with a Proxy.
-	nodePorts := ports.PopIntSlice(3)
+	nodePorts := helpers.NewPortSlice(3)
 	proxyReverseTunnelPort, proxyWebPort, proxySSHPort := nodePorts[0], nodePorts[1], nodePorts[2]
-	proxyConfig := ProxyConfig{
+	proxyConfig := helpers.ProxyConfig{
 		Name:              "cluster-main-proxy",
 		SSHPort:           proxySSHPort,
 		WebPort:           proxyWebPort,
@@ -3177,13 +3120,13 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 
 	// Wait for active tunnel connections to be established.
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 1)
-	waitForActiveTunnelConnections(t, proxyTunnel, Site, 1)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 1)
+	waitForActiveTunnelConnections(t, proxyTunnel, helpers.Site, 1)
 
 	// Execute the connection via first proxy.
-	cfg := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cfg := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    "cluster-main-node",
 	}
 	output, err := runCommand(t, main, []string{"echo", "hello world"}, cfg, 1)
@@ -3194,9 +3137,9 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 	// tried 10 times with 250 millisecond delay between each attempt to allow
 	// the discovery request to be received and the connection added to the agent
 	// pool.
-	cfgProxy := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cfgProxy := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    "cluster-main-node",
 		Proxy:   &proxyConfig,
 	}
@@ -3207,7 +3150,7 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 
 	// Remove second proxy from LB.
 	require.NoError(t, lb.RemoveBackend(*proxyTwoBackend))
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 1)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 1)
 
 	// Requests going via main proxy will succeed. Requests going via second
 	// proxy will fail.
@@ -3219,8 +3162,8 @@ func testDiscoveryNode(t *testing.T, suite *integrationTestSuite) {
 
 	// Add second proxy to LB, both should have a connection.
 	lb.AddBackend(*proxyTwoBackend)
-	waitForActiveTunnelConnections(t, main.Tunnel, Site, 1)
-	waitForActiveTunnelConnections(t, proxyTunnel, Site, 1)
+	waitForActiveTunnelConnections(t, main.Tunnel, helpers.Site, 1)
+	waitForActiveTunnelConnections(t, proxyTunnel, helpers.Site, 1)
 
 	// Requests going via both proxies will succeed.
 	output, err = runCommand(t, main, []string{"echo", "hello world"}, cfg, 1)
@@ -3265,7 +3208,7 @@ func waitForActivePeerProxyConnections(t *testing.T, tunnel reversetunnel.Server
 
 // waitForProxyCount waits a set time for the proxy count in clusterName to
 // reach some value.
-func waitForProxyCount(t *TeleInstance, clusterName string, count int) error {
+func waitForProxyCount(t *helpers.TeleInstance, clusterName string, count int) error {
 	var counts map[string]int
 	start := time.Now()
 	for time.Since(start) < 17*time.Second {
@@ -3281,7 +3224,7 @@ func waitForProxyCount(t *TeleInstance, clusterName string, count int) error {
 }
 
 // waitForNodeCount waits for a certain number of nodes to show up in the remote site.
-func waitForNodeCount(ctx context.Context, t *TeleInstance, clusterName string, count int) error {
+func waitForNodeCount(ctx context.Context, t *helpers.TeleInstance, clusterName string, count int) error {
 	const (
 		deadline     = time.Second * 30
 		iterWaitTime = time.Second
@@ -3391,7 +3334,7 @@ func testExternalClient(t *testing.T, suite *integrationTestSuite) {
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
 			// Create a Teleport instance with auth, proxy, and node.
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 					Mode: tt.inRecordLocation,
 				})
@@ -3409,18 +3352,18 @@ func testExternalClient(t *testing.T, suite *integrationTestSuite) {
 
 				return t, nil, nil, tconf
 			}
-			teleport := suite.newTeleportWithConfig(makeConfig())
+			teleport := suite.NewTeleportWithConfig(makeConfig())
 			defer teleport.StopAll()
 
 			// Generate certificates for the user simulating login.
-			creds, err := GenerateUserCreds(UserCredsRequest{
+			creds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{
 				Process:  teleport.Process,
-				Username: suite.me.Username,
+				Username: suite.Me.Username,
 			})
 			require.NoError(t, err)
 
 			// Start (and defer close) a agent that runs during this integration test.
-			teleAgent, socketDirPath, socketPath, err := createAgent(suite.me, creds.Key.Priv, creds.Key.Cert)
+			teleAgent, socketDirPath, socketPath, err := createAgent(suite.Me, creds.Key.Priv, creds.Key.Cert)
 			require.NoError(t, err)
 			defer closeAgent(teleAgent, socketDirPath)
 
@@ -3487,7 +3430,7 @@ func testControlMaster(t *testing.T, suite *integrationTestSuite) {
 		controlPath := filepath.Join(controlDir, "control-path")
 
 		// Create a Teleport instance with auth, proxy, and node.
-		makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+		makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 			recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 				Mode: tt.inRecordLocation,
 			})
@@ -3505,18 +3448,18 @@ func testControlMaster(t *testing.T, suite *integrationTestSuite) {
 
 			return t, nil, nil, tconf
 		}
-		teleport := suite.newTeleportWithConfig(makeConfig())
+		teleport := suite.NewTeleportWithConfig(makeConfig())
 		defer teleport.StopAll()
 
 		// Generate certificates for the user simulating login.
-		creds, err := GenerateUserCreds(UserCredsRequest{
+		creds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{
 			Process:  teleport.Process,
-			Username: suite.me.Username,
+			Username: suite.Me.Username,
 		})
 		require.NoError(t, err)
 
 		// Start (and defer close) a agent that runs during this integration test.
-		teleAgent, socketDirPath, socketPath, err := createAgent(suite.me, creds.Key.Priv, creds.Key.Cert)
+		teleAgent, socketDirPath, socketPath, err := createAgent(suite.Me, creds.Key.Priv, creds.Key.Cert)
 		require.NoError(t, err)
 		defer closeAgent(teleAgent, socketDirPath)
 
@@ -3579,19 +3522,19 @@ func testProxyHostKeyCheck(t *testing.T, suite *integrationTestSuite) {
 
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
-			hostSigner, err := ssh.ParsePrivateKey(suite.priv)
+			hostSigner, err := ssh.ParsePrivateKey(suite.Priv)
 			require.NoError(t, err)
 
 			// start a ssh server that presents a host key instead of a certificate
-			nodePort := ports.PopInt()
-			sshNode, err := newDiscardServer(Host, nodePort, hostSigner)
+			nodePort := helpers.NewPortValue()
+			sshNode, err := helpers.NewDiscardServer(Host, nodePort, hostSigner)
 			require.NoError(t, err)
 			err = sshNode.Start()
 			require.NoError(t, err)
 			defer sshNode.Stop()
 
 			// create a teleport instance with auth, proxy, and node
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 					Mode:                types.RecordAtProxy,
 					ProxyChecksHostKeys: types.NewBoolOption(tt.inHostKeyCheck),
@@ -3608,13 +3551,13 @@ func testProxyHostKeyCheck(t *testing.T, suite *integrationTestSuite) {
 
 				return t, nil, nil, tconf
 			}
-			teleport := suite.newTeleportWithConfig(makeConfig())
+			teleport := suite.NewTeleportWithConfig(makeConfig())
 			defer teleport.StopAll()
 
 			// create a teleport client and exec a command
-			clientConfig := ClientConfig{
-				Login:        suite.me.Username,
-				Cluster:      Site,
+			clientConfig := helpers.ClientConfig{
+				Login:        suite.Me.Username,
+				Cluster:      helpers.Site,
 				Host:         Host,
 				Port:         nodePort,
 				ForwardAgent: true,
@@ -3640,7 +3583,7 @@ func testAuditOff(t *testing.T, suite *integrationTestSuite) {
 	var err error
 
 	// create a teleport instance with auth, proxy, and node
-	makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+	makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 		recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 			Mode: types.RecordOff,
 		})
@@ -3658,11 +3601,11 @@ func testAuditOff(t *testing.T, suite *integrationTestSuite) {
 
 		return t, nil, nil, tconf
 	}
-	teleport := suite.newTeleportWithConfig(makeConfig())
+	teleport := suite.NewTeleportWithConfig(makeConfig())
 	defer teleport.StopAll()
 
 	// get access to a authClient for the cluster
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	// should have no sessions in it to start with
@@ -3674,9 +3617,9 @@ func testAuditOff(t *testing.T, suite *integrationTestSuite) {
 
 	myTerm := NewTerminal(250)
 	go func() {
-		cl, err := teleport.NewClient(ClientConfig{
-			Login:   suite.me.Username,
-			Cluster: Site,
+		cl, err := teleport.NewClient(helpers.ClientConfig{
+			Login:   suite.Me.Username,
+			Cluster: helpers.Site,
 			Host:    Host,
 			Port:    teleport.GetPortSSHInt(),
 		})
@@ -3701,7 +3644,7 @@ func testAuditOff(t *testing.T, suite *integrationTestSuite) {
 		require.NoError(t, err)
 	}
 	// make sure it's us who joined! :)
-	require.Equal(t, suite.me.Username, session.Parties[0].User)
+	require.Equal(t, suite.Me.Username, session.Parties[0].User)
 
 	// lets type "echo hi" followed by "enter" and then "exit" + "enter":
 	myTerm.Type("\aecho hi\n\r\aexit\n\r\a")
@@ -3823,7 +3766,7 @@ func testPAM(t *testing.T, suite *integrationTestSuite) {
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
 			// Create a teleport instance with auth, proxy, and node.
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				tconf := suite.defaultServiceConfig()
 				tconf.Auth.Enabled = true
 
@@ -3839,7 +3782,7 @@ func testPAM(t *testing.T, suite *integrationTestSuite) {
 
 				return t, nil, nil, tconf
 			}
-			teleport := suite.newTeleportWithConfig(makeConfig())
+			teleport := suite.NewTeleportWithConfig(makeConfig())
 			defer teleport.StopAll()
 
 			termSession := NewTerminal(250)
@@ -3847,9 +3790,9 @@ func testPAM(t *testing.T, suite *integrationTestSuite) {
 			// Create an interactive session and write something to the terminal.
 			ctx, cancel := context.WithCancel(context.Background())
 			go func() {
-				cl, err := teleport.NewClient(ClientConfig{
-					Login:   suite.me.Username,
-					Cluster: Site,
+				cl, err := teleport.NewClient(helpers.ClientConfig{
+					Login:   suite.Me.Username,
+					Cluster: helpers.Site,
 					Host:    Host,
 					Port:    teleport.GetPortSSHInt(),
 				})
@@ -3895,10 +3838,10 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	teleport := suite.newTeleportInstance()
+	teleport := suite.NewTeleportInstance()
 	defer teleport.StopAll()
 
-	logins := []string{suite.me.Username}
+	logins := []string{suite.Me.Username}
 	for _, login := range logins {
 		teleport.AddUser(login, []string{login})
 	}
@@ -3909,14 +3852,14 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 
 	// Enable Kubernetes service to test issue where the `KubernetesReady` event was not properly propagated
 	// and in the case where Kube service was enabled cert rotation flow was broken.
-	enableKubernetesService(t, config)
+	helpers.EnableKubernetesService(t, config)
 
 	serviceC := make(chan *service.TeleportProcess, 20)
 
 	runErrCh := make(chan error, 1)
 	go func() {
 		runErrCh <- service.Run(ctx, *config, func(cfg *service.Config) (service.Process, error) {
-			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&disabledIMDSClient{}))
+			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&helpers.DisabledIMDSClient{}))
 			if err == nil {
 				serviceC <- svc
 			}
@@ -3935,18 +3878,18 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 		id, err := svc.GetIdentity(types.RoleNode)
 		require.NoError(t, err)
 		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname)
-		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname+"."+Site)
-		require.Contains(t, id.Cert.ValidPrincipals, HostID)
-		require.Contains(t, id.Cert.ValidPrincipals, HostID+"."+Site)
+		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname+"."+helpers.Site)
+		require.Contains(t, id.Cert.ValidPrincipals, helpers.HostID)
+		require.Contains(t, id.Cert.ValidPrincipals, helpers.HostID+"."+helpers.Site)
 	}
 	checkSSHPrincipals(svc)
 
 	// Setup user in the cluster
-	err = SetupUser(svc, suite.me.Username, nil)
+	err = helpers.SetupUser(svc, suite.Me.Username, nil)
 	require.NoError(t, err)
 
 	// capture credentials before reload started to simulate old client
-	initialCreds, err := GenerateUserCreds(UserCredsRequest{Process: svc, Username: suite.me.Username})
+	initialCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{Process: svc, Username: suite.Me.Username})
 	require.NoError(t, err)
 
 	t.Logf("Service started. Setting rotation state to %v", types.RotationPhaseUpdateClients)
@@ -3958,7 +3901,7 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	})
 	require.NoError(t, err)
 
-	hostCA, err := svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: Site}, false)
+	hostCA, err := svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: helpers.Site}, false)
 	require.NoError(t, err)
 	t.Logf("Cert authority: %v", auth.CertAuthorityInfo(hostCA))
 
@@ -3977,9 +3920,9 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	svc, err = waitForReload(serviceC, svc)
 	require.NoError(t, err)
 
-	cfg := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	cfg := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Loopback,
 		Port:    teleport.GetPortSSHInt(),
 	}
@@ -4000,7 +3943,7 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	})
 	require.NoError(t, err)
 
-	hostCA, err = svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: Site}, false)
+	hostCA, err = svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: helpers.Site}, false)
 	require.NoError(t, err)
 	t.Logf("Cert authority: %v", auth.CertAuthorityInfo(hostCA))
 
@@ -4009,7 +3952,7 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 
 	// new credentials will work from this phase to others
-	newCreds, err := GenerateUserCreds(UserCredsRequest{Process: svc, Username: suite.me.Username})
+	newCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{Process: svc, Username: suite.Me.Username})
 	require.NoError(t, err)
 
 	clt, err = teleport.NewClientWithCreds(cfg, *newCreds)
@@ -4029,7 +3972,7 @@ func testRotateSuccess(t *testing.T, suite *integrationTestSuite) {
 	})
 	require.NoError(t, err)
 
-	hostCA, err = svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: Site}, false)
+	hostCA, err = svc.GetAuthServer().GetCertAuthority(ctx, types.CertAuthID{Type: types.HostCA, DomainName: helpers.Site}, false)
 	require.NoError(t, err)
 	t.Logf("Cert authority: %v", auth.CertAuthorityInfo(hostCA))
 
@@ -4066,9 +4009,9 @@ func testRotateRollback(t *testing.T, s *integrationTestSuite) {
 	defer cancel()
 
 	tconf := s.rotationConfig(true)
-	teleport := s.newTeleportInstance()
+	teleport := s.NewTeleportInstance()
 	defer teleport.StopAll()
-	logins := []string{s.me.Username}
+	logins := []string{s.Me.Username}
 	for _, login := range logins {
 		teleport.AddUser(login, []string{login})
 	}
@@ -4080,7 +4023,7 @@ func testRotateRollback(t *testing.T, s *integrationTestSuite) {
 	runErrCh := make(chan error, 1)
 	go func() {
 		runErrCh <- service.Run(ctx, *config, func(cfg *service.Config) (service.Process, error) {
-			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&disabledIMDSClient{}))
+			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&helpers.DisabledIMDSClient{}))
 			if err == nil {
 				serviceC <- svc
 			}
@@ -4099,18 +4042,18 @@ func testRotateRollback(t *testing.T, s *integrationTestSuite) {
 		id, err := svc.GetIdentity(types.RoleNode)
 		require.NoError(t, err)
 		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname)
-		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname+"."+Site)
-		require.Contains(t, id.Cert.ValidPrincipals, HostID)
-		require.Contains(t, id.Cert.ValidPrincipals, HostID+"."+Site)
+		require.Contains(t, id.Cert.ValidPrincipals, svc.Config.Hostname+"."+helpers.Site)
+		require.Contains(t, id.Cert.ValidPrincipals, helpers.HostID)
+		require.Contains(t, id.Cert.ValidPrincipals, helpers.HostID+"."+helpers.Site)
 	}
 	checkSSHPrincipals(svc)
 
 	// Setup user in the cluster
-	err = SetupUser(svc, s.me.Username, nil)
+	err = helpers.SetupUser(svc, s.Me.Username, nil)
 	require.NoError(t, err)
 
 	// capture credentials before reload started to simulate old client
-	initialCreds, err := GenerateUserCreds(UserCredsRequest{Process: svc, Username: s.me.Username})
+	initialCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{Process: svc, Username: s.Me.Username})
 	require.NoError(t, err)
 
 	t.Logf("Service started. Setting rotation state to %q.", types.RotationPhaseInit)
@@ -4138,9 +4081,9 @@ func testRotateRollback(t *testing.T, s *integrationTestSuite) {
 	svc, err = waitForReload(serviceC, svc)
 	require.NoError(t, err)
 
-	cfg := ClientConfig{
-		Login:   s.me.Username,
-		Cluster: Site,
+	cfg := helpers.ClientConfig{
+		Login:   s.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Loopback,
 		Port:    teleport.GetPortSSHInt(),
 	}
@@ -4213,7 +4156,7 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	main := suite.newNamedTeleportInstance(t, clusterMain)
 	aux := suite.newNamedTeleportInstance(t, clusterAux)
 
-	logins := []string{suite.me.Username}
+	logins := []string{suite.Me.Username}
 	for _, login := range logins {
 		main.AddUser(login, []string{login})
 	}
@@ -4224,7 +4167,7 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	runErrCh := make(chan error, 1)
 	go func() {
 		runErrCh <- service.Run(ctx, *config, func(cfg *service.Config) (service.Process, error) {
-			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&disabledIMDSClient{}))
+			svc, err := service.NewTeleport(cfg, service.WithIMDSClient(&helpers.DisabledIMDSClient{}))
 			if err == nil {
 				serviceC <- svc
 			}
@@ -4239,12 +4182,12 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	mainDevs := "main-devs"
 	role, err := types.NewRoleV3(mainDevs, types.RoleSpecV5{
 		Allow: types.RoleConditions{
-			Logins: []string{suite.me.Username},
+			Logins: []string{suite.Me.Username},
 		},
 	})
 	require.NoError(t, err)
 
-	err = SetupUser(svc, suite.me.Username, []types.Role{role})
+	err = helpers.SetupUser(svc, suite.Me.Username, []types.Role{role})
 	require.NoError(t, err)
 
 	// create auxiliary cluster and setup trust
@@ -4257,7 +4200,7 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	auxDevs := "aux-devs"
 	role, err = types.NewRoleV3(auxDevs, types.RoleSpecV5{
 		Allow: types.RoleConditions{
-			Logins: []string{suite.me.Username},
+			Logins: []string{suite.Me.Username},
 		},
 	})
 	require.NoError(t, err)
@@ -4280,15 +4223,15 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	waitForTunnelConnections(t, svc.GetAuthServer(), aux.Secrets.SiteName, 1)
 
 	// capture credentials before reload has started to simulate old client
-	initialCreds, err := GenerateUserCreds(UserCredsRequest{
+	initialCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{
 		Process:  svc,
-		Username: suite.me.Username,
+		Username: suite.Me.Username,
 	})
 	require.NoError(t, err)
 
 	// credentials should work
-	cfg := ClientConfig{
-		Login:   suite.me.Username,
+	cfg := helpers.ClientConfig{
+		Login:   suite.Me.Username,
 		Host:    Loopback,
 		Cluster: clusterAux,
 		Port:    aux.GetPortSSHInt(),
@@ -4368,7 +4311,7 @@ func testRotateTrustedClusters(t *testing.T, suite *integrationTestSuite) {
 	waitForPhase(types.RotationPhaseUpdateServers)
 
 	// new credentials will work from this phase to others
-	newCreds, err := GenerateUserCreds(UserCredsRequest{Process: svc, Username: suite.me.Username})
+	newCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{Process: svc, Username: suite.Me.Username})
 	require.NoError(t, err)
 
 	clt, err = main.NewClientWithCreds(cfg, *newCreds)
@@ -4528,7 +4471,7 @@ func testWindowChange(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleport(t, nil, true)
 	defer teleport.StopAll()
 
-	site := teleport.GetSiteAPI(Site)
+	site := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, site)
 
 	personA := NewTerminal(250)
@@ -4536,9 +4479,9 @@ func testWindowChange(t *testing.T, suite *integrationTestSuite) {
 
 	// openSession will open a new session on a server.
 	openSession := func() {
-		cl, err := teleport.NewClient(ClientConfig{
-			Login:   suite.me.Username,
-			Cluster: Site,
+		cl, err := teleport.NewClient(helpers.ClientConfig{
+			Login:   suite.Me.Username,
+			Cluster: helpers.Site,
 			Host:    Host,
 			Port:    teleport.GetPortSSHInt(),
 		})
@@ -4562,9 +4505,9 @@ func testWindowChange(t *testing.T, suite *integrationTestSuite) {
 		require.NoError(t, err)
 		sessionID := string(sessions[0].ID)
 
-		cl, err := teleport.NewClient(ClientConfig{
-			Login:   suite.me.Username,
-			Cluster: Site,
+		cl, err := teleport.NewClient(helpers.ClientConfig{
+			Login:   suite.Me.Username,
+			Cluster: helpers.Site,
 			Host:    Host,
 			Port:    teleport.GetPortSSHInt(),
 		})
@@ -4646,7 +4589,7 @@ func testList(t *testing.T, suite *integrationTestSuite) {
 	defer tr.Stop()
 
 	// Create and start a Teleport cluster with auth, proxy, and node.
-	makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+	makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 		recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 			Mode: types.RecordOff,
 		})
@@ -4666,11 +4609,11 @@ func testList(t *testing.T, suite *integrationTestSuite) {
 
 		return t, nil, nil, tconf
 	}
-	teleport := suite.newTeleportWithConfig(makeConfig())
+	teleport := suite.NewTeleportWithConfig(makeConfig())
 	defer teleport.StopAll()
 
 	// Create and start a Teleport node.
-	nodeSSHPort := ports.PopInt()
+	nodeSSHPort := helpers.NewPortValue()
 	nodeConfig := func() *service.Config {
 		tconf := suite.defaultServiceConfig()
 		tconf.Hostname = "server-02"
@@ -4686,7 +4629,7 @@ func testList(t *testing.T, suite *integrationTestSuite) {
 	require.NoError(t, err)
 
 	// Get an auth client to the cluster.
-	clt := teleport.GetSiteAPI(Site)
+	clt := teleport.GetSiteAPI(helpers.Site)
 	require.NotNil(t, clt)
 
 	// Wait 10 seconds for both nodes to show up to make sure they both have
@@ -4753,15 +4696,15 @@ func testList(t *testing.T, suite *integrationTestSuite) {
 			require.NoError(t, err)
 
 			// Create user, role, and generate credentials.
-			err = SetupUser(teleport.Process, tt.inLogin, []types.Role{role})
+			err = helpers.SetupUser(teleport.Process, tt.inLogin, []types.Role{role})
 			require.NoError(t, err)
-			initialCreds, err := GenerateUserCreds(UserCredsRequest{Process: teleport.Process, Username: tt.inLogin})
+			initialCreds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{Process: teleport.Process, Username: tt.inLogin})
 			require.NoError(t, err)
 
 			// Create a Teleport client.
-			cfg := ClientConfig{
+			cfg := helpers.ClientConfig{
 				Login:   tt.inLogin,
-				Cluster: Site,
+				Cluster: helpers.Site,
 				Port:    teleport.GetPortSSHInt(),
 			}
 			userClt, err := teleport.NewClientWithCreds(cfg, *initialCreds)
@@ -4813,7 +4756,7 @@ func testCmdLabels(t *testing.T, suite *integrationTestSuite) {
 
 		return tconf
 	}
-	teleport := suite.newTeleportWithConfig(t, nil, nil, makeConfig())
+	teleport := suite.NewTeleportWithConfig(t, nil, nil, makeConfig())
 	defer teleport.StopAll()
 
 	// Create and start a reversetunnel node.
@@ -4861,9 +4804,9 @@ func testCmdLabels(t *testing.T, suite *integrationTestSuite) {
 
 	for _, tt := range tts {
 		t.Run(tt.desc, func(t *testing.T) {
-			cfg := ClientConfig{
-				Login:   suite.me.Username,
-				Cluster: Site,
+			cfg := helpers.ClientConfig{
+				Login:   suite.Me.Username,
+				Cluster: helpers.Site,
 				Labels:  tt.labels,
 			}
 
@@ -4888,9 +4831,9 @@ func testDataTransfer(t *testing.T, suite *integrationTestSuite) {
 	defer main.StopAll()
 
 	// Create a client to the above Teleport cluster.
-	clientConfig := ClientConfig{
-		Login:   suite.me.Username,
-		Cluster: Site,
+	clientConfig := helpers.ClientConfig{
+		Login:   suite.Me.Username,
+		Cluster: helpers.Site,
 		Host:    Host,
 		Port:    main.GetPortSSHInt(),
 	}
@@ -4962,7 +4905,7 @@ func testBPFInteractive(t *testing.T, suite *integrationTestSuite) {
 			dir := t.TempDir()
 
 			// Create and start a Teleport cluster.
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 					Mode: tt.inSessionRecording,
 				})
@@ -4990,7 +4933,7 @@ func testBPFInteractive(t *testing.T, suite *integrationTestSuite) {
 				}
 				return t, nil, nil, tconf
 			}
-			main := suite.newTeleportWithConfig(makeConfig())
+			main := suite.NewTeleportWithConfig(makeConfig())
 			defer main.StopAll()
 
 			// Create a client terminal and context to signal when the client is done
@@ -4999,9 +4942,9 @@ func testBPFInteractive(t *testing.T, suite *integrationTestSuite) {
 			doneContext, doneCancel := context.WithCancel(context.Background())
 
 			func() {
-				client, err := main.NewClient(ClientConfig{
-					Login:   suite.me.Username,
-					Cluster: Site,
+				client, err := main.NewClient(helpers.ClientConfig{
+					Login:   suite.Me.Username,
+					Cluster: helpers.Site,
 					Host:    Host,
 					Port:    main.GetPortSSHInt(),
 				})
@@ -5090,7 +5033,7 @@ func testBPFExec(t *testing.T, suite *integrationTestSuite) {
 			dir := t.TempDir()
 
 			// Create and start a Teleport cluster.
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 					Mode: tt.inSessionRecording,
 				})
@@ -5118,13 +5061,13 @@ func testBPFExec(t *testing.T, suite *integrationTestSuite) {
 				}
 				return t, nil, nil, tconf
 			}
-			main := suite.newTeleportWithConfig(makeConfig())
+			main := suite.NewTeleportWithConfig(makeConfig())
 			defer main.StopAll()
 
 			// Create a client to the above Teleport cluster.
-			clientConfig := ClientConfig{
-				Login:   suite.me.Username,
-				Cluster: Site,
+			clientConfig := helpers.ClientConfig{
+				Login:   suite.Me.Username,
+				Cluster: helpers.Site,
 				Host:    Host,
 				Port:    main.GetPortSSHInt(),
 			}
@@ -5216,7 +5159,7 @@ func testSSHExitCode(t *testing.T, suite *integrationTestSuite) {
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
 			// Create and start a Teleport cluster.
-			makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+			makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 				// Create default config.
 				tconf := suite.defaultServiceConfig()
 
@@ -5234,16 +5177,16 @@ func testSSHExitCode(t *testing.T, suite *integrationTestSuite) {
 				tconf.SSH.Enabled = true
 				return t, nil, nil, tconf
 			}
-			main := suite.newTeleportWithConfig(makeConfig())
+			main := suite.NewTeleportWithConfig(makeConfig())
 			t.Cleanup(func() { main.StopAll() })
 
 			// context to signal when the client is done with the terminal.
 			doneContext, doneCancel := context.WithTimeout(context.Background(), time.Second*10)
 			defer doneCancel()
 
-			cli, err := main.NewClient(ClientConfig{
-				Login:       suite.me.Username,
-				Cluster:     Site,
+			cli, err := main.NewClient(helpers.ClientConfig{
+				Login:       suite.Me.Username,
+				Cluster:     helpers.Site,
 				Host:        Host,
 				Port:        main.GetPortSSHInt(),
 				Interactive: tt.interactive,
@@ -5293,7 +5236,7 @@ func testBPFSessionDifferentiation(t *testing.T, suite *integrationTestSuite) {
 	dir := t.TempDir()
 
 	// Create and start a Teleport cluster.
-	makeConfig := func() (*testing.T, []string, []*InstanceSecrets, *service.Config) {
+	makeConfig := func() (*testing.T, []string, []*helpers.InstanceSecrets, *service.Config) {
 		recConfig, err := types.NewSessionRecordingConfigFromConfigFile(types.SessionRecordingConfigSpecV2{
 			Mode: types.RecordAtNode,
 		})
@@ -5319,7 +5262,7 @@ func testBPFSessionDifferentiation(t *testing.T, suite *integrationTestSuite) {
 		tconf.SSH.BPF.CgroupPath = dir
 		return t, nil, nil, tconf
 	}
-	main := suite.newTeleportWithConfig(makeConfig())
+	main := suite.NewTeleportWithConfig(makeConfig())
 	defer main.StopAll()
 
 	// Create two client terminals and channel to signal when the clients are
@@ -5330,9 +5273,9 @@ func testBPFSessionDifferentiation(t *testing.T, suite *integrationTestSuite) {
 
 	// Open a terminal and type "ls" into both and exit.
 	writeTerm := func(term *Terminal) {
-		client, err := main.NewClient(ClientConfig{
-			Login:   suite.me.Username,
-			Cluster: Site,
+		client, err := main.NewClient(helpers.ClientConfig{
+			Login:   suite.Me.Username,
+			Cluster: helpers.Site,
 			Host:    Host,
 			Port:    main.GetPortSSHInt(),
 		})
@@ -5438,9 +5381,9 @@ func testExecEvents(t *testing.T, suite *integrationTestSuite) {
 	for _, tt := range execTests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Create client for each test in grid tests
-			clientConfig := ClientConfig{
-				Login:       suite.me.Username,
-				Cluster:     Site,
+			clientConfig := helpers.ClientConfig{
+				Login:       suite.Me.Username,
+				Cluster:     helpers.Site,
 				Host:        Host,
 				Port:        main.GetPortSSHInt(),
 				Interactive: tt.isInteractive,
@@ -5491,7 +5434,7 @@ func testSessionStartContainsAccessRequest(t *testing.T, suite *integrationTestS
 		Options: types.RoleOptions{},
 		Allow: types.RoleConditions{
 			Logins: []string{
-				suite.me.Username,
+				suite.Me.Username,
 			},
 			Request: &types.AccessRequestConditions{
 				Roles: []string{requestedRoleName},
@@ -5503,7 +5446,7 @@ func testSessionStartContainsAccessRequest(t *testing.T, suite *integrationTestS
 	err = authServer.UpsertRole(ctx, userRole)
 	require.NoError(t, err)
 
-	user, err := types.NewUser(suite.me.Username)
+	user, err := types.NewUser(suite.Me.Username)
 	user.AddRole(userRole.GetName())
 	require.NoError(t, err)
 
@@ -5534,7 +5477,7 @@ func testSessionStartContainsAccessRequest(t *testing.T, suite *integrationTestS
 
 	WaitForResource(t, watcher, user.GetKind(), user.GetName())
 
-	req, err := services.NewAccessRequest(suite.me.Username, requestedRole.GetMetadata().Name)
+	req, err := services.NewAccessRequest(suite.Me.Username, requestedRole.GetMetadata().Name)
 	require.NoError(t, err)
 
 	accessRequestID := req.GetName()
@@ -5550,9 +5493,9 @@ func testSessionStartContainsAccessRequest(t *testing.T, suite *integrationTestS
 
 	WaitForResource(t, watcher, req.GetKind(), req.GetName())
 
-	clientConfig := ClientConfig{
-		Login:       suite.me.Username,
-		Cluster:     Site,
+	clientConfig := helpers.ClientConfig{
+		Login:       suite.Me.Username,
+		Cluster:     helpers.Site,
 		Host:        Host,
 		Port:        main.GetPortSSHInt(),
 		Interactive: false,
@@ -5596,7 +5539,7 @@ func WaitForResource(t *testing.T, watcher types.Watcher, kind, name string) {
 }
 
 // findEventInLog polls the event log looking for an event of a particular type.
-func findEventInLog(t *TeleInstance, eventName string) (events.EventFields, error) {
+func findEventInLog(t *helpers.TeleInstance, eventName string) (events.EventFields, error) {
 	for i := 0; i < 10; i++ {
 		eventFields, err := eventsInLog(t.Config.DataDir+"/log/events.log", eventName)
 		if err != nil {
@@ -5620,7 +5563,7 @@ func findEventInLog(t *TeleInstance, eventName string) (events.EventFields, erro
 }
 
 // findCommandEventInLog polls the event log looking for an event of a particular type.
-func findCommandEventInLog(t *TeleInstance, eventName string, programName string) (events.EventFields, error) {
+func findCommandEventInLog(t *helpers.TeleInstance, eventName string, programName string) (events.EventFields, error) {
 	for i := 0; i < 10; i++ {
 		eventFields, err := eventsInLog(t.Config.DataDir+"/log/events.log", eventName)
 		if err != nil {
@@ -5674,7 +5617,7 @@ func eventsInLog(path string, eventName string) ([]events.EventFields, error) {
 }
 
 // runCommandWithCertReissue runs an SSH command and generates certificates for the user
-func runCommandWithCertReissue(t *testing.T, instance *TeleInstance, cmd []string, reissueParams client.ReissueParams, cachePolicy client.CertCachePolicy, cfg ClientConfig) error {
+func runCommandWithCertReissue(t *testing.T, instance *helpers.TeleInstance, cmd []string, reissueParams client.ReissueParams, cachePolicy client.CertCachePolicy, cfg helpers.ClientConfig) error {
 	tc, err := instance.NewClient(cfg)
 	if err != nil {
 		return trace.Wrap(err)
@@ -5699,7 +5642,7 @@ func runCommandWithCertReissue(t *testing.T, instance *TeleInstance, cmd []strin
 // connected to proxy of the passed in instance, runs the command, and returns
 // the result. If multiple attempts are requested, a 250 millisecond delay is
 // added between them before giving up.
-func runCommand(t *testing.T, instance *TeleInstance, cmd []string, cfg ClientConfig, attempts int) (string, error) {
+func runCommand(t *testing.T, instance *helpers.TeleInstance, cmd []string, cfg helpers.ClientConfig, attempts int) (string, error) {
 	tc, err := instance.NewClient(cfg)
 	if err != nil {
 		return "", trace.Wrap(err)
@@ -5730,42 +5673,26 @@ func runCommand(t *testing.T, instance *TeleInstance, cmd []string, cfg ClientCo
 	return output.String(), nil
 }
 
-func (s *integrationTestSuite) newTeleportInstance() *TeleInstance {
-	return NewInstance(s.defaultInstanceConfig())
-}
+type InstanceConfigOption func(config *helpers.InstanceConfig)
 
-func (s *integrationTestSuite) defaultInstanceConfig() InstanceConfig {
-	return InstanceConfig{
-		ClusterName: Site,
-		HostID:      HostID,
-		NodeName:    Host,
-		Priv:        s.priv,
-		Pub:         s.pub,
-		log:         s.log,
-		Ports:       standardPortSetup(),
-	}
-}
-
-type InstanceConfigOption func(config *InstanceConfig)
-
-func (s *integrationTestSuite) newNamedTeleportInstance(t *testing.T, clusterName string, opts ...InstanceConfigOption) *TeleInstance {
-	cfg := InstanceConfig{
+func (s *integrationTestSuite) newNamedTeleportInstance(t *testing.T, clusterName string, opts ...InstanceConfigOption) *helpers.TeleInstance {
+	cfg := helpers.InstanceConfig{
 		ClusterName: clusterName,
-		HostID:      HostID,
+		HostID:      helpers.HostID,
 		NodeName:    Host,
-		Priv:        s.priv,
-		Pub:         s.pub,
-		log:         utils.WrapLogger(s.log.WithField("cluster", clusterName)),
-		Ports:       standardPortSetup(),
+		Priv:        s.Priv,
+		Pub:         s.Pub,
+		Log:         utils.WrapLogger(s.Log.WithField("cluster", clusterName)),
+		Ports:       helpers.StandardPortSetup(),
 	}
 	for _, opt := range opts {
 		opt(&cfg)
 	}
-	return NewInstance(cfg)
+	return helpers.NewInstance(cfg)
 }
 
 func WithNodeName(nodeName string) InstanceConfigOption {
-	return func(config *InstanceConfig) {
+	return func(config *helpers.InstanceConfig) {
 		config.NodeName = nodeName
 	}
 }
@@ -5773,7 +5700,7 @@ func WithNodeName(nodeName string) InstanceConfigOption {
 func (s *integrationTestSuite) defaultServiceConfig() *service.Config {
 	cfg := service.MakeDefaultConfig()
 	cfg.Console = nil
-	cfg.Log = s.log
+	cfg.Log = s.Log
 	cfg.CircuitBreakerConfig = breaker.NoopBreakerConfig()
 	return cfg
 }
@@ -5849,13 +5776,13 @@ func TestWebProxyInsecure(t *testing.T) {
 	privateKey, publicKey, err := testauthority.New().GenerateKeyPair()
 	require.NoError(t, err)
 
-	rc := NewInstance(InstanceConfig{
+	rc := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: "example.com",
 		HostID:      uuid.New().String(),
 		NodeName:    Host,
 		Priv:        privateKey,
 		Pub:         publicKey,
-		log:         utils.NewLoggerForTests(),
+		Log:         utils.NewLoggerForTests(),
 	})
 
 	rcConf := service.MakeDefaultConfig()
@@ -5894,23 +5821,23 @@ func TestTraitsPropagation(t *testing.T) {
 	require.NoError(t, err)
 
 	// Create root cluster.
-	rc := NewInstance(InstanceConfig{
+	rc := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: "root.example.com",
 		HostID:      uuid.New().String(),
 		NodeName:    Host,
 		Priv:        privateKey,
 		Pub:         publicKey,
-		log:         log,
+		Log:         log,
 	})
 
 	// Create leaf cluster.
-	lc := NewInstance(InstanceConfig{
+	lc := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: "leaf.example.com",
 		HostID:      uuid.New().String(),
 		NodeName:    Host,
 		Priv:        privateKey,
 		Pub:         publicKey,
-		log:         log,
+		Log:         log,
 	})
 
 	// Make root cluster config.
@@ -5979,7 +5906,7 @@ func TestTraitsPropagation(t *testing.T) {
 	require.NoError(t, err)
 
 	// Run command in root.
-	outputRoot, err := runCommand(t, rc, []string{"echo", "hello root"}, ClientConfig{
+	outputRoot, err := runCommand(t, rc, []string{"echo", "hello root"}, helpers.ClientConfig{
 		Login:   me.Username,
 		Cluster: "root.example.com",
 		Host:    Loopback,
@@ -5989,7 +5916,7 @@ func TestTraitsPropagation(t *testing.T) {
 	require.Equal(t, "hello root", strings.TrimSpace(outputRoot))
 
 	// Run command in leaf.
-	outputLeaf, err := runCommand(t, rc, []string{"echo", "hello leaf"}, ClientConfig{
+	outputLeaf, err := runCommand(t, rc, []string{"echo", "hello leaf"}, helpers.ClientConfig{
 		Login:   me.Username,
 		Cluster: "leaf.example.com",
 		Host:    Loopback,
@@ -6006,7 +5933,7 @@ func testSessionStreaming(t *testing.T, suite *integrationTestSuite) {
 	teleport := suite.newTeleport(t, nil, true)
 	defer teleport.StopAll()
 
-	api := teleport.GetSiteAPI(Site)
+	api := teleport.GetSiteAPI(helpers.Site)
 	uploadStream, err := api.CreateAuditStream(ctx, sessionID)
 	require.Nil(t, err)
 
@@ -6156,9 +6083,9 @@ func testKubeAgentFiltering(t *testing.T, suite *integrationTestSuite) {
 			err = adminSite.CreateUser(ctx, testCase.user)
 			require.NoError(t, err)
 
-			cl, err := teleport.NewClient(ClientConfig{
+			cl, err := teleport.NewClient(helpers.ClientConfig{
 				Login:   testCase.user.GetName(),
-				Cluster: Site,
+				Cluster: helpers.Site,
 				Host:    Host,
 				Port:    teleport.GetPortSSHInt(),
 			})
@@ -6167,7 +6094,7 @@ func testKubeAgentFiltering(t *testing.T, suite *integrationTestSuite) {
 			proxy, err := cl.ConnectToProxy(ctx)
 			require.NoError(t, err)
 
-			userSite, err := proxy.ConnectToCluster(ctx, Site)
+			userSite, err := proxy.ConnectToCluster(ctx, helpers.Site)
 			require.NoError(t, err)
 
 			services, err := userSite.GetKubeServices(ctx)
@@ -6177,30 +6104,30 @@ func testKubeAgentFiltering(t *testing.T, suite *integrationTestSuite) {
 	}
 }
 
-func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraServices func(*testing.T, *TeleInstance, *TeleInstance)) *client.TeleportClient {
+func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraServices func(*testing.T, *helpers.TeleInstance, *helpers.TeleInstance)) *client.TeleportClient {
 	ctx := context.Background()
-	username := suite.me.Username
+	username := suite.Me.Username
 	name := "test"
 	rootName := fmt.Sprintf("root-%s", name)
 	leafName := fmt.Sprintf("leaf-%s", name)
 
 	// Create root and leaf clusters.
-	root := NewInstance(InstanceConfig{
+	root := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: rootName,
-		HostID:      HostID,
+		HostID:      helpers.HostID,
 		NodeName:    Host,
-		Priv:        suite.priv,
-		Pub:         suite.pub,
-		log:         suite.log,
+		Priv:        suite.Priv,
+		Pub:         suite.Pub,
+		Log:         suite.Log,
 		Ports:       standardPortsOrMuxSetup(false),
 	})
-	leaf := NewInstance(InstanceConfig{
+	leaf := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: leafName,
-		HostID:      HostID,
+		HostID:      helpers.HostID,
 		NodeName:    Host,
-		Priv:        suite.priv,
-		Pub:         suite.pub,
-		log:         suite.log,
+		Priv:        suite.Priv,
+		Pub:         suite.Pub,
+		Log:         suite.Log,
 		Ports:       standardPortsOrMuxSetup(false),
 	})
 
@@ -6212,7 +6139,7 @@ func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraSe
 	require.NoError(t, err)
 	root.AddUserWithRole(username, role)
 
-	makeConfig := func() (*testing.T, []*InstanceSecrets, *service.Config) {
+	makeConfig := func() (*testing.T, []*helpers.InstanceSecrets, *service.Config) {
 		tconf := suite.defaultServiceConfig()
 		tconf.Proxy.DisableWebService = false
 		tconf.Proxy.DisableWebInterface = true
@@ -6247,13 +6174,13 @@ func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraSe
 	tryCreateTrustedCluster(t, leaf.Process.GetAuthServer(), trustedCluster)
 	waitForTunnelConnections(t, root.Process.GetAuthServer(), leafName, 1)
 
-	rootSSHPort := ports.PopInt()
-	rootProxySSHPort := ports.PopInt()
-	rootProxyWebPort := ports.PopInt()
+	rootSSHPort := helpers.NewPortValue()
+	rootProxySSHPort := helpers.NewPortValue()
+	rootProxyWebPort := helpers.NewPortValue()
 	require.NoError(t, root.StartNodeAndProxy("root-zero", rootSSHPort, rootProxyWebPort, rootProxySSHPort))
-	leafSSHPort := ports.PopInt()
-	leafProxySSHPort := ports.PopInt()
-	leafProxyWebPort := ports.PopInt()
+	leafSSHPort := helpers.NewPortValue()
+	leafProxySSHPort := helpers.NewPortValue()
+	leafProxyWebPort := helpers.NewPortValue()
 	require.NoError(t, leaf.StartNodeAndProxy("leaf-zero", leafSSHPort, leafProxyWebPort, leafProxySSHPort))
 
 	// Add any extra services.
@@ -6265,14 +6192,14 @@ func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraSe
 	require.Eventually(t, waitForClusters(leaf.Tunnel, 1), 10*time.Second, 1*time.Second)
 
 	// Create client.
-	creds, err := GenerateUserCreds(UserCredsRequest{
+	creds, err := helpers.GenerateUserCreds(helpers.UserCredsRequest{
 		Process:        root.Process,
 		Username:       username,
 		RouteToCluster: rootName,
 	})
 	require.NoError(t, err)
 
-	tc, err := root.NewClientWithCreds(ClientConfig{
+	tc, err := root.NewClientWithCreds(helpers.ClientConfig{
 		Login:   username,
 		Cluster: rootName,
 		Host:    Loopback,
@@ -6290,12 +6217,12 @@ func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraSe
 }
 
 func testListResourcesAcrossClusters(t *testing.T, suite *integrationTestSuite) {
-	tc := createTrustedClusterPair(t, suite, func(t *testing.T, root, leaf *TeleInstance) {
+	tc := createTrustedClusterPair(t, suite, func(t *testing.T, root, leaf *helpers.TeleInstance) {
 		rootNodes := []string{"root-one", "root-two"}
 		leafNodes := []string{"leaf-one", "leaf-two"}
 
 		// Start a Teleport node that has SSH, Apps, Databases, and Kubernetes.
-		startNode := func(name string, i *TeleInstance) {
+		startNode := func(name string, i *helpers.TeleInstance) {
 			conf := suite.defaultServiceConfig()
 			conf.Auth.Enabled = false
 			conf.Proxy.Enabled = false
@@ -6313,7 +6240,7 @@ func testListResourcesAcrossClusters(t *testing.T, suite *integrationTestSuite) 
 				Enabled: true,
 			}
 			conf.SSH.Addr = utils.NetAddr{
-				Addr: fmt.Sprintf("%s:%d", Host, ports.PopInt()),
+				Addr: fmt.Sprintf("%s:%d", Host, helpers.NewPortValue()),
 			}
 			conf.Proxy.Enabled = false
 
@@ -6335,7 +6262,7 @@ func testListResourcesAcrossClusters(t *testing.T, suite *integrationTestSuite) 
 			}
 
 			conf.Kube.KubeconfigPath = filepath.Join(conf.DataDir, "kube_config")
-			require.NoError(t, enableKube(conf, name))
+			require.NoError(t, helpers.EnableKube(conf, name))
 			conf.Kube.ListenAddr = nil
 			process, err := service.NewTeleport(conf)
 			require.NoError(t, err)
@@ -6351,7 +6278,7 @@ func testListResourcesAcrossClusters(t *testing.T, suite *integrationTestSuite) 
 				service.TeleportReadyEvent,
 			}
 
-			receivedEvents, err := startAndWait(process, expectedEvents)
+			receivedEvents, err := helpers.StartAndWait(process, expectedEvents)
 			require.NoError(t, err)
 			log.Debugf("Teleport Kube Server (in instance %v) started: %v/%v events received.",
 				i.Secrets.SiteName, len(expectedEvents), len(receivedEvents))
