+++ integration/port_forwarding_test.go
@@ -28,6 +28,7 @@ import (
 
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/integration/helpers"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/client"
 	"github.com/gravitational/teleport/lib/session"
@@ -105,10 +106,10 @@ func testPortForwarding(t *testing.T, suite *integrationTestSuite) {
 			cfg.SSH.Enabled = true
 			cfg.SSH.AllowTCPForwarding = tt.portForwardingAllowed
 
-			teleport := suite.newTeleportWithConfig(t, nil, nil, cfg)
+			teleport := suite.NewTeleportWithConfig(t, nil, nil, cfg)
 			defer teleport.StopAll()
 
-			site := teleport.GetSiteAPI(Site)
+			site := teleport.GetSiteAPI(helpers.Site)
 
 			// ...and a running dummy server
 			remoteSvr := httptest.NewServer(http.HandlerFunc(
@@ -120,14 +121,14 @@ func testPortForwarding(t *testing.T, suite *integrationTestSuite) {
 
 			// ... and a client connection that was launched with port
 			// forwarding enabled to that dummy server
-			localPort := ports.PopInt()
+			localPort := helpers.NewPortValue()
 			remotePort, err := extractPort(remoteSvr)
 			require.NoError(t, err)
 
 			nodeSSHPort := teleport.GetPortSSHInt()
-			cl, err := teleport.NewClient(ClientConfig{
-				Login:   suite.me.Username,
-				Cluster: Site,
+			cl, err := teleport.NewClient(helpers.ClientConfig{
+				Login:   suite.Me.Username,
+				Cluster: helpers.Site,
 				Host:    Host,
 				Port:    nodeSSHPort,
 			})
