+++ integration/proxy_helpers_test.go
@@ -29,6 +29,7 @@ import (
 	"time"
 
 	"github.com/gravitational/teleport/api/breaker"
+	"github.com/gravitational/teleport/integration/helpers"
 	"github.com/gravitational/trace"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -54,8 +55,8 @@ import (
 )
 
 type ProxySuite struct {
-	root *TeleInstance
-	leaf *TeleInstance
+	root *helpers.TeleInstance
+	leaf *helpers.TeleInstance
 }
 
 type proxySuiteOptions struct {
@@ -68,11 +69,11 @@ type proxySuiteOptions struct {
 	rootClusterNodeName string
 	leafClusterNodeName string
 
-	rootClusterPorts *InstancePorts
-	leafClusterPorts *InstancePorts
+	rootClusterPorts *helpers.InstancePorts
+	leafClusterPorts *helpers.InstancePorts
 
-	rootTrustedSecretFunc func(suite *ProxySuite) []*InstanceSecrets
-	leafTrustedFunc       func(suite *ProxySuite) []*InstanceSecrets
+	rootTrustedSecretFunc func(suite *ProxySuite) []*helpers.InstanceSecrets
+	leafTrustedFunc       func(suite *ProxySuite) []*helpers.InstanceSecrets
 
 	rootClusterRoles      []types.Role
 	leafClusterRoles      []types.Role
@@ -85,29 +86,29 @@ func newProxySuite(t *testing.T, opts ...proxySuiteOptionsFunc) *ProxySuite {
 	options := proxySuiteOptions{
 		rootClusterNodeName: Host,
 		leafClusterNodeName: Host,
-		rootClusterPorts:    singleProxyPortSetup(),
-		leafClusterPorts:    singleProxyPortSetup(),
+		rootClusterPorts:    helpers.SingleProxyPortSetup(),
+		leafClusterPorts:    helpers.SingleProxyPortSetup(),
 	}
 	for _, opt := range opts {
 		opt(&options)
 	}
 
-	rc := NewInstance(InstanceConfig{
+	rc := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: "root.example.com",
 		HostID:      uuid.New().String(),
 		NodeName:    options.rootClusterNodeName,
-		log:         utils.NewLoggerForTests(),
+		Log:         utils.NewLoggerForTests(),
 		Ports:       options.rootClusterPorts,
 	})
 
 	// Create leaf cluster.
-	lc := NewInstance(InstanceConfig{
+	lc := helpers.NewInstance(helpers.InstanceConfig{
 		ClusterName: "leaf.example.com",
 		HostID:      uuid.New().String(),
 		NodeName:    options.leafClusterNodeName,
 		Priv:        rc.Secrets.PrivKey,
 		Pub:         rc.Secrets.PubKey,
-		log:         utils.NewLoggerForTests(),
+		Log:         utils.NewLoggerForTests(),
 		Ports:       options.leafClusterPorts,
 	})
 	suite := &ProxySuite{
@@ -201,7 +202,7 @@ func (p *ProxySuite) addNodeToLeafCluster(t *testing.T, tunnelNodeHostname strin
 	require.NoError(t, err)
 }
 
-func (p *ProxySuite) mustConnectToClusterAndRunSSHCommand(t *testing.T, config ClientConfig) {
+func (p *ProxySuite) mustConnectToClusterAndRunSSHCommand(t *testing.T, config helpers.ClientConfig) {
 	const (
 		deadline         = time.Second * 5
 		nextIterWaitTime = time.Millisecond * 100
@@ -262,10 +263,10 @@ func withRootClusterConfig(fn func(suite *ProxySuite) *service.Config, configMod
 
 func withRootAndLeafTrustedClusterReset() proxySuiteOptionsFunc {
 	return func(options *proxySuiteOptions) {
-		options.rootTrustedSecretFunc = func(suite *ProxySuite) []*InstanceSecrets {
+		options.rootTrustedSecretFunc = func(suite *ProxySuite) []*helpers.InstanceSecrets {
 			return nil
 		}
-		options.leafTrustedFunc = func(suite *ProxySuite) []*InstanceSecrets {
+		options.leafTrustedFunc = func(suite *ProxySuite) []*helpers.InstanceSecrets {
 			return nil
 		}
 	}
@@ -283,13 +284,13 @@ func withLeafClusterNodeName(nodeName string) proxySuiteOptionsFunc {
 	}
 }
 
-func withRootClusterPorts(ports *InstancePorts) proxySuiteOptionsFunc {
+func withRootClusterPorts(ports *helpers.InstancePorts) proxySuiteOptionsFunc {
 	return func(options *proxySuiteOptions) {
 		options.rootClusterPorts = ports
 	}
 }
 
-func withLeafClusterPorts(ports *InstancePorts) proxySuiteOptionsFunc {
+func withLeafClusterPorts(ports *helpers.InstancePorts) proxySuiteOptionsFunc {
 	return func(options *proxySuiteOptions) {
 		options.leafClusterPorts = ports
 	}
@@ -516,7 +517,7 @@ func makeNodeConfig(nodeName, authAddr string) *service.Config {
 	return nodeConfig
 }
 
-func mustCreateUserIdentityFile(t *testing.T, tc *TeleInstance, username string, ttl time.Duration) string {
+func mustCreateUserIdentityFile(t *testing.T, tc *helpers.TeleInstance, username string, ttl time.Duration) string {
 	key, err := libclient.NewKey()
 	require.NoError(t, err)
 	key.ClusterName = tc.Secrets.SiteName
